Changed autosave code.
